Replace mobile+OTP subscription auth with email+password; security and cleanup pass
Free hobby project, no billing.

- Remove the OTP/carrier-billing subscription system entirely.
- Add email+password registration, login, and password reset for both
  patients and nutritionists.
- Close an open redirect in the post-login `next` param (accepted
  protocol-relative URLs like //evil.com).
- Fix an unhandled duplicate-email race in both registration paths, and a
  login timing side-channel that could reveal whether an email is
  registered.
- Add Vite build tooling for CSS/JS; gitignore node_modules/ properly.
- Trim dead code found in an over-engineering pass: unused Request/Logger
  methods, an unused Validator rule, and a duplicate exception-handling
  branch in the front controller.

+++ app/Controllers/HealthController.php
@@ -14,7 +14,7 @@ final class HealthController extends Controller
     public function index(Request $request): Response
     {
         try {
-            $ok = Db::value('SELECT 1') === 1;
+            $ok = (int) Db::value('SELECT 1') === 1;
         } catch (\Throwable) {
             $ok = false;
         }
